Rename package fileutil to files
