+++ kadai1/exaream/imgconv/converter.go
@@ -14,7 +14,7 @@ import (
 	"reflect"
 	"strings"
 
-	"assignment/fileutil"
+	"assignment/files"
 	"assignment/slices"
 
 	"golang.org/x/image/bmp"
@@ -86,7 +86,7 @@ func ValidateArgs() error {
 		return fmt.Errorf("the %v must be selected from: %v", argDstExt, extListStr)
 	}
 
-	srcDir, err := fileutil.OpenFile(*SrcDir)
+	srcDir, err := files.OpenFile(*SrcDir)
 	if err != nil {
 		if os.IsNotExist(err) {
 			return fmt.Errorf("the %v does not exist: %w", argSrcDir, err)
@@ -135,7 +135,7 @@ func (conv *Converter) Run() error {
 			return nil
 		}
 		// Skip the process if an acquired extension and a source extension are different
-		if ext := fileutil.GetFormattedFileExt(srcFilePath); ext != conv.srcExt {
+		if ext := files.GetFormattedFileExt(srcFilePath); ext != conv.srcExt {
 			return nil
 		}
 		// Get an absolute path of a destination directory
@@ -189,7 +189,7 @@ func (conv *Converter) getRelPathSrcDir(srcFilePath string) (string, error) {
 
 // Get a destination file path
 func (conv *Converter) getDstFilePath(srcFilePath string, dstDir string) string {
-	dstFileName := fileutil.GetFileStem(srcFilePath) + conv.dstExt
+	dstFileName := files.GetFileStem(srcFilePath) + conv.dstExt
 	return filepath.Join(dstDir, dstFileName)
 }
 
@@ -230,7 +230,7 @@ func (conv *Converter) encode(dstImage io.Writer, srcImage image.Image) error {
 // Get a source image
 // (Use named return values to return an error from the inside of defer)
 func getSrcImage(srcFilePath string) (srcImage image.Image, err error) {
-	srcFile, err := fileutil.OpenFile(srcFilePath)
+	srcFile, err := files.OpenFile(srcFilePath)
 	if err != nil {
 		return nil, err
 	}
